+++ demo/scheme/src/jvmMain/kotlin/Main.kt
@@ -8,18 +8,18 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
-import center.sciprog.maps.features.FeatureGroup
-import center.sciprog.maps.features.ViewConfig
-import center.sciprog.maps.features.ViewPoint
-import center.sciprog.maps.features.color
-import center.sciprog.maps.scheme.*
-import center.sciprog.maps.svg.FeatureStateSnapshot
-import center.sciprog.maps.svg.exportToSvg
-import center.sciprog.maps.svg.snapshot
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.isActive
 import kotlinx.coroutines.launch
 import space.kscience.kmath.geometry.Angle
+import space.kscience.maps.features.FeatureGroup
+import space.kscience.maps.features.ViewConfig
+import space.kscience.maps.features.ViewPoint
+import space.kscience.maps.features.color
+import space.kscience.maps.scheme.*
+import space.kscience.maps.svg.FeatureStateSnapshot
+import space.kscience.maps.svg.exportToSvg
+import space.kscience.maps.svg.snapshot
 import java.awt.Desktop
 import java.nio.file.Files
 
@@ -29,7 +29,7 @@ fun App() {
     MaterialTheme {
         val scope = rememberCoroutineScope()
 
-        val schemeFeaturesState: FeatureGroup<XY> = FeatureGroup.remember(XYCoordinateSpace) {
+        val features: FeatureGroup<XY> = FeatureGroup.remember(XYCoordinateSpace) {
             background(1600f, 1200f) { painterResource("middle-earth.jpg") }
             circle(410.52737 to 868.7676).color(Color.Blue)
             text(410.52737 to 868.7676, "Shire").color(Color.Blue)
@@ -53,7 +53,7 @@ fun App() {
         }
 
         val initialViewPoint: ViewPoint<XY> = remember {
-            schemeFeaturesState.getBoundingBox(1f)?.computeViewPoint() ?: XYViewPoint(XY(0f, 0f))
+            features.getBoundingBox(1f)?.computeViewPoint() ?: XYViewPoint(XY(0f, 0f))
         }
 
         var viewPoint: ViewPoint<XY> by remember { mutableStateOf(initialViewPoint) }
@@ -61,7 +61,7 @@ fun App() {
         var snapshot: FeatureStateSnapshot<XY>? by remember { mutableStateOf(null) }
 
         if (snapshot == null) {
-            snapshot = schemeFeaturesState.snapshot()
+            snapshot = features.snapshot()
         }
 
         ContextMenuArea(
@@ -78,7 +78,7 @@ fun App() {
                 )
             }
         ) {
-            val mapState: XYViewScope = XYViewScope.remember(
+            val mapState: XYCanvasState = XYCanvasState.remember(
                 ViewConfig(
                     onClick = { _, click ->
                         println("${click.focus.x}, ${click.focus.y}")
@@ -90,7 +90,7 @@ fun App() {
 
             SchemeView(
                 mapState,
-                schemeFeaturesState,
+                features,
             )
         }