+++ CHANGELOG.md
@@ -3,13 +3,27 @@
 ## Unreleased
 
 ### Added
+- `alpha` extension for feature attribute builder
 
 ### Changed
+- avoid drawing features with VisibleAttribute false
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+- Add alpha attribute comprehension for all standard features.
 
 ### Security
+
+## 0.3.0 - 2024-06-04
+
+### Changed
+
+- Package changed to `space.kscience`
+- Kotlin 2.0
+
+### Fixed
+
+- Use of generated resources for Wasm